Runtime Error Message: Line 8: ValueError: negative number cannot be raised to a fractional power Last executed input: -6

+++ learnpythonthehardway/largesrt-palindrome-product.py
@@ -43,8 +43,8 @@ class Solution(object):
             s = str(i) + str(i)[::-1]
             j = upper
             while j * j > int(s):
-                if int(s) % j == 0:
-                    return int(s) % 1337
+                if long(s) % j == 0:
+                    return long(s) % 1337
                 j -= 1
         return 9
